tests: Move ExternalInterfaceTestProvider to its own mod
